+++ Model/Method/Alma4X.php
@@ -17,21 +17,21 @@
 namespace HiPay\FullserviceMagento\Model\Method;
 
 /**
- * Klarna invoice Model payment method
+ * Alma 4X payment method
  *
  * @author    Kassim Belghait <kassim@sirateck.com>
  * @copyright Copyright (c) 2016 - HiPay
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
  * @link      https://github.com/hipay/hipay-fullservice-sdk-magento2
  */
-class Klarna extends HostedMethod
+class Alma4X extends AbstractMethodAPI
 {
-    public const HIPAY_METHOD_CODE = 'hipay_klarna';
+    public const HIPAY_METHOD_CODE = 'hipay_alma4X';
 
     /**
      * @var string
      */
-    protected static $_technicalCode = 'klarnainvoice';
+    protected static $_technicalCode = 'alma-4x';
 
     /**
      * @var string