+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.20.0
+
+- **Add** : Added support for new payment means :
+  - Alma 3x
+  - Alma 4x
+- **Add** : Removed Split Payment method. Existing split payments are still processed normally.
+- **Add** : Removed Hosted Page v1 option
+- **Add** : Removed unsupported Klarna Invoice payment method
+- **Fix** : Fixed issues with deprecations in PHP 8.2
+
 ## 1.19.1
 
 - **Fix** : Fixed issues with deprecations in PHP 8.2